+++ sdk/test/index.ts
@@ -1,1 +1,2 @@
+export * from './server-setup';
 export * from './test-utils';
