+++ sdk/test/test-utils.tsx
@@ -6,8 +6,6 @@ import { ThemeProvider } from '@0xsequence/design-system';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import type { RenderOptions } from '@testing-library/react';
 import { renderHook, render as rtlRender } from '@testing-library/react';
-import { HttpResponse, http as mswHttp } from 'msw';
-import { setupServer } from 'msw/node';
 import type { ReactElement } from 'react';
 import {
 	type Client,
@@ -18,30 +16,9 @@ import {
 import { mainnet as wagmiMainet, polygon as wagmiPolygon } from 'viem/chains';
 import { type Config, createConfig, http, WagmiProvider } from 'wagmi';
 import { mock } from 'wagmi/connectors';
-import { handlers as marketplaceConfigHandlers } from '../src/react/_internal/api/__mocks__/builder.msw';
-import { handlers as indexerHandlers } from '../src/react/_internal/api/__mocks__/indexer.msw';
-import { laosHandlers } from '../src/react/_internal/api/__mocks__/laos.msw';
-import { handlers as marketplaceHandlers } from '../src/react/_internal/api/__mocks__/marketplace.msw';
-import { handlers as metadataHandlers } from '../src/react/_internal/api/__mocks__/metadata.msw';
 import { TEST_ACCOUNTS, TEST_CHAIN, TEST_PRIVATE_KEYS } from './const';
 
-const tickHandler = mswHttp.post(
-	'https://nodes.sequence.app/rpc/Databeat/Tick',
-	() => {
-		return HttpResponse.json({});
-	},
-);
-
-export const server = setupServer(
-	...marketplaceHandlers,
-	...metadataHandlers,
-	...indexerHandlers,
-	...marketplaceConfigHandlers,
-	...laosHandlers,
-	tickHandler,
-);
-
-const createTestQueryClient = () =>
+export const createTestQueryClient = () =>
 	new QueryClient({
 		defaultOptions: {
 			queries: {
@@ -92,6 +69,13 @@ const wagmiConfigObj = {
 
 export const wagmiConfig = createConfig(wagmiConfigObj);
 
+export const sequenceConnectConfig = createSequenceConnectConfig('universal', {
+	projectAccessKey: 'test',
+	chainIds: [1, 137],
+	defaultChainId: 1,
+	appName: 'Demo Dapp',
+});
+
 function mockWaas() {
 	return new Proxy(mockConnector, {
 		apply(target, thisArg, args) {
@@ -140,13 +124,6 @@ function renderWithClient(ui: ReactElement, options?: Options) {
 				: // if testing non-sequence connector, use the default config
 					wagmiConfig;
 	}
-
-	const sequenceConnectConfig = createSequenceConnectConfig('universal', {
-		projectAccessKey: 'test',
-		chainIds: [1, 137],
-		defaultChainId: 1,
-		appName: 'Demo Dapp',
-	});
 
 	const Wrapper = ({ children }: { children: React.ReactNode }) => (
 		<WagmiProvider config={config}>
